dynamic button based for different status update/delete leave record amend OT function

+++ LAPS/src/main/java/SA50/T6/WadCA/LAPS/service/OvertimeServiceImpl.java
@@ -14,7 +14,7 @@ public class OvertimeServiceImpl implements OvertimeService{
 	@Override
 	public float FindCompensationLeaveAwarded(int overtime) {
 		int daysAwarded = overtime/4;
-		float unitsAwarded = daysAwarded/2;
+		float unitsAwarded = (float)daysAwarded/2;
 		return unitsAwarded;
 	}
